+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.0.9 ##
 
-* Updated to support new version of UserEcho API
+* Updated to support new version of UserEcho API.
+* Comments counts on the topic list fixed, title in the comments list fixed.
+* Fixed status pane height in the topic view when official answer text is empty.